requirements.txt
anthropic>=0.39.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-pool>=0.6.0
chromadb>=0.5.0

# Resolution chain
feedparser>=6.0.0
rapidfuzz>=3.0.0
google-genai>=1.0.0

# Synthesis output + Google Drive upload
python-docx>=1.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
